Add Int boxing/unboxing to the Nim VM.

+++ vmnim/values.nim
@@ -5,6 +5,6 @@ type
     tpe*: Type  # TODO (vm): How can we make this an IntType for IntValues? Maybe the variant objects aren't a good idea after all?
 
   IntValue* = ref object of Value
-    value*: int
+    value*: int64
 
-proc new_int*(value: int): IntValue = IntValue(value: value, tpe: types.int)
+proc new_int*(value: int64): IntValue = IntValue(value: value, tpe: types.int)
